==> code/step3-check.Rmd <==
---
title: "step3-check"
author: "ylchen"
date: "2021年3月22日"
output:
  rmdformats::downcute:
    self_contained: true
---

## 0. 加载R包

```{r,eval=FALSE}
rm(list = ls())   
options(stringsAsFactors = F)
library("ChAMP")
library("minfi")
library("wateRmelon")
require(GEOquery)
require(Biobase)
library("FactoMineR")#画主成分分析图需要加载这两个包
library("factoextra")
```

## 1. 加载数据

```{r,collapse=TRUE}
load(file = '../data/step1-output.Rdata')
beta.m=myLoad$beta
group_list=myLoad$pd$type

table(group_list)
dim(beta.m)
beta.m[1:4,1:4]
```

## 2. 画图

```{r,include=FALSE}
  # 使用 wateRmelon 进行 归一化
  library("wateRmelon")
```

### 2.1 rawBox

```{r}
  beta.m=beta.m[rowMeans(beta.m)>0.005,]
  #dim(beta.m)
  #pdf(file="rawBox.pdf")
  boxplot(beta.m,col = "blue",xaxt = "n",outline = F)
```

### 2.2 normalBox

```{r}
  
  beta.m = betaqn(beta.m)
  #pdf(file="normalBox.pdf")
  boxplot(beta.m,col = "red",xaxt = "n",outline = F)
```

### 2.3 densityBeanPlot

```{r}
  
  # 然后进行简单的QC 
  #pdf(file="densityBeanPlot.pdf")
  #par(oma=c(2,10,2,2))
  densityBeanPlot(beta.m, sampGroups = group_list)
```

### 2.4 mdsPlot

```{r}  
  #pdf(file="mdsPlot.pdf")
  mdsPlot(beta.m, numPositions = 1000, sampGroups = group_list )

 
```

```{r,include=FALSE}
  library("FactoMineR")#画主成分分析图需要加载这两个包
  library("factoextra")
  library('pheatmap')
```

### 2.5 标准三张图
 * PCA
 * pheatmap
 * cor
 
```{r}
# 标准三张图 

  dat=t(beta.m)[,sample(1:nrow(beta.m),10000)]
  #dat[1:4,1:4] 
  # library("FactoMineR")#画主成分分析图需要加载这两个包
  # library("factoextra")  
  # 因为甲基化芯片是450K或者850K，几十万行的甲基化位点，所以PCA不会太快
  dat.pca <- PCA(dat , graph = FALSE) 
  fviz_pca_ind(dat.pca,
               geom.ind = "point", # show points only (nbut not "text")
               col.ind = group_list, # color by groups
               # palette = c("#00AFBB", "#E7B800"),
               addEllipses = TRUE, # Concentration ellipses
               legend.title = "Groups")
  #ggsave('../pic/all_samples_PCA.png')
  
  dat=beta.m
  #dat[1:4,1:4] 
  cg=names(tail(sort(apply(dat,1,sd)),1000))#apply按行（'1'是按行取，'2'是按列取）取每一行的方差，从小到大排序，取最大的1000个
  #library(pheatmap)
  #pheatmap(dat[cg,],show_colnames =F,show_rownames = F) #对那些提取出来的1000个基因所在的每一行取出，组合起来为一个新的表达矩阵
  n=t(scale(t(dat[cg,]))) # 'scale'可以对log-ratio数值进行归一化
  n[n>2]=2 
  n[n< -2]= -2
  n[1:4,1:4]
  #pheatmap(n,show_colnames =F,show_rownames = F)
  ac=data.frame(group=group_list)
  rownames(ac)=colnames(n) 
  pheatmap(n,show_colnames =F,show_rownames = F,
           annotation_col=ac)
  # #pheatmap(n,show_colnames =F,show_rownames = F,
  #          annotation_col=ac,filename = '../pic/heatmap_top1000_sd.png')
  # dev.off()
  
  exprSet=beta.m
  #pheatmap::pheatmap(cor(exprSet)) 
  # 组内的样本的相似性应该是要高于组间的！
  colD=data.frame(group_list=group_list)
  rownames(colD)=colnames(exprSet)
  pheatmap::pheatmap(cor(exprSet),
                     annotation_col = colD,
                     show_rownames = F)
  # pheatmap::pheatmap(cor(exprSet),
  #                    annotation_col = colD,
  #                    show_rownames = F,
  #                    filename = '../pic/cor_all.png')
  # dev.off() 
  exprSet=exprSet[names(sort(apply(exprSet, 1,mad),decreasing = T)[1:500]),]
  #dim(exprSet)
  # M=cor(log2(exprSet+1)) 
  M=cor(exprSet)
  pheatmap::pheatmap(M,annotation_col = colD)
  # #pheatmap::pheatmap(M,
  #                    show_rownames = F,
  #                    annotation_col = colD,
  #                    filename = '../pic/cor_top500.png')
  # #dev.off() 
```
